% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createpalette.R
\name{add_palette}
\alias{add_palette}
\title{Internal function for adding or updating a palette to the data file}
\usage{
add_palette(jpegfile, palette)
}
\arguments{
\item{jpegfile}{link to the file - use a local file if possible}

\item{palette}{name for the palette}
}
\description{
Already existing palettes will be updated, new palettes will be added to the `ga_pal_cont` object in the right format.
}
\examples{
gretchenalbrecht:::add_palette("pictures/gretchen-albrecht-changes.jpg", "changes")
}
