% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{ga_pal}
\alias{ga_pal}
\title{gretchenalbrecht palette with ramped colours}
\usage{
ga_pal(palette = "rocker", alpha = 1, reverse = FALSE)
}
\arguments{
\item{palette}{Choose from 'gretchenalbrecht_palettes' list}

\item{alpha}{transparency}

\item{reverse}{If TRUE, the direction of the colours is reversed.}
}
\description{
gretchenalbrecht palette with ramped colours
}
\examples{
library(scales)
show_col(ga_pal()(10))

}
