% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{gretchenalbrecht_palettes}
\alias{gretchenalbrecht_palettes}
\title{Gretchen Albrecht Colour Palettes}
\format{An object of class \code{list} of length 8.}
\usage{
gretchenalbrecht_palettes
}
\description{
A collection of colour palettes inspired by the paintings of NZ abstract expressionist Gretchen Albrecht:
rocker
black_plain
flood_tide
last_rays
red_sky_golden_cloud
winged_spill
pink_cloud
winter_light
}
\examples{

# Make an x-y plot using the Namatjira palette
library(tidyverse)
df <- data.frame(x = rnorm(100, 0, 20),
          y = rnorm(100, 0, 20),
          cl = sample(letters[1:8], 100, replace=TRUE))
ggplot(df, aes(x, y, colour=cl, shape=cl)) +
  geom_point(size=4) + scale_colour_gretchenalbrecht() +
  theme_bw() + theme(aspect.ratio=1)

# Make a histogram using the McCrea Collins Street palette
ggplot(df, aes(x, fill=cl)) + geom_histogram() +
  scale_fill_gretchenalbrecht(palette="pink_cloud")

}
\keyword{datasets}
