% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nz_cart_census}
\alias{nz_cart_census}
\title{Cartogram coordinates and 2013 census data}
\format{A data frame with 6812 rows with the following variables:
\itemize{
  \item long: Longitude.
  \item lat: Latitude
  \item order, hole, piece, id, group: components to draw polygons
  \item Name: Name of the region
  \item REGC: id of region
  \item PropUnemploymentBenefit2013: Proportion of population on unemployment benefit
  \item Prop65AndOlder_2013: Proportion older than 65
  \item PropDoctorate2013: Proportion with doctorate
  \item MedianIncome2013: Median income
  }}
\usage{
nz_cart_census
}
\description{
Data subsetted from Peter Ellis's nzelect package
and uses his cartogram view of NZ based on population
}
\keyword{datasets}
