% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_gretchenalbrecht}
\alias{scale_color_gretchenalbrecht}
\alias{scale_colour_gretchenalbrecht}
\title{Setup colour palette for ggplot2}
\usage{
scale_color_gretchenalbrecht(..., palette = "rocker", discrete = TRUE,
  alpha = 1, reverse = FALSE)

scale_colour_gretchenalbrecht(..., palette = "rocker", discrete = TRUE,
  alpha = 1, reverse = FALSE)
}
\arguments{
\item{...}{additional arguments to pass to scale_color_gradientn}

\item{palette}{Choose from 'gretchenalbrecht_palettes' list}

\item{discrete}{whether to use a discrete colour palette}

\item{alpha}{transparency}

\item{reverse}{logical, Reverse the order of the colours?}
}
\description{
Setup colour palette for ggplot2
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(mpg, wt)) +
  geom_point(aes(colour = factor(cyl))) +
  scale_colour_gretchenalbrecht(palette="rocker")
ggplot(mtcars, aes(mpg, wt)) +
  geom_point(aes(colour = hp)) +
  scale_colour_gretchenalbrecht(palette="rocker", discrete = FALSE)
ggplot(data = mpg) +
  geom_point(mapping = aes(x = displ, y = hwy, color = class)) +
  scale_colour_gretchenalbrecht(palette="rocker")
ggplot(diamonds) + geom_bar(aes(x = cut, fill = clarity)) +
  scale_fill_gretchenalbrecht()
library(ggthemes)
data(nz_cart_census)
ggplot(nz_cart_census, aes(x=long, y=lat, group=group,
 fill=MedianIncome2013)) +
 scale_fill_gretchenalbrecht(palette="winter_light",
                             discrete=FALSE) +
 geom_polygon() + theme_map() + theme(legend.position="right")
}
