% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createpalette.R
\name{get_pal}
\alias{get_pal}
\title{Get k colors from a named color palette}
\usage{
get_pal(K, palette = "changes")
}
\arguments{
\item{K}{number of colors}

\item{palette}{name of a palette}
}
\value{
dataframe with columns k (from 1 to K) and column col
}
\description{
Get k colors from a named color palette
}
\examples{
# colours from a luminous shade:
ls <- get_pal(5, "luminous shade")
ls \%>\% ggplot() + 
   geom_tile(aes(x=1, y = 1, fill=col)) + 
   scale_fill_identity() + facet_wrap(~k)
}
