% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{gretchenalbrecht_palettes}
\alias{gretchenalbrecht_palettes}
\title{Gretchen Albrecht Colour Palettes}
\format{An object of class \code{list} of length 8.}
\usage{
gretchenalbrecht_palettes
}
\description{
A collection of colour palettes inspired by the paintings of NZ abstract expressionist Gretchen Albrecht:
rocker
black_plain
flood_tide
last_rays
red_sky_golden_cloud
winged_spill
pink_cloud
winter_light
}
\examples{

# Print out the palettes available
gretchenalbrecht_palettes

# Make an x-y plot using the rocker palette
library(tidyverse)
data(diamonds)
diamonds_small <- diamonds \%>\% sample_n(1000)
ggplot(diamonds_small,
  aes(x=carat, price, colour=cut, shape=cut)) +
  geom_point(size=4, alpha=0.5) +
  scale_colour_gretchenalbrecht(palette="rocker") +
  theme_bw() + theme(aspect.ratio=1)

# Make a histogram using the pink_cloud palette
ggplot(diamonds_small, aes(x=price, fill=cut)) + geom_histogram() +
  scale_fill_gretchenalbrecht(palette="pink_cloud") + theme_bw()

}
\keyword{datasets}
