% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createpalette.R
\name{make_palette}
\alias{make_palette}
\title{Create a palette of (almost) arbitrary size from given image}
\usage{
make_palette(jpgfile)
}
\arguments{
\item{jpgfile}{name of the image from which the color palette is to be extracted}
}
\value{
cluster object
}
\description{
Create a palette of (almost) arbitrary size from given image
}
\examples{
ls <- make_palette("pictures/gretchen-albrecht-luminous-shade-811.jpg")
ls$name <- "luminous shade"
# now add to internal data object to make it accessible for future use
}
